[Feat] 게시글 하나, 여러개 받아오는 loader 구현
