+++ src/App.js
@@ -4,8 +4,11 @@ import RootLayout from '@pages/RootLayout';
 import ErrorPage from '@pages/Error';
 
 import HomeRootLayout from '@pages/HomeRootLayout';
-import HomePage from '@pages/home/Home';
-import PostDetailPage from '@pages/home/PostDetail';
+import HomePage, { loader as postsLoader } from '@pages/home/Home';
+import PostDetailPage, {
+  loader as postDetailLoader,
+} from '@pages/home/PostDetail';
+import ProfilePage, { loader as profileLoader } from '@pages/home/Profile';
 import NotificationPage from '@pages/home/Notification';
 
 import AuthenticationPage from '@pages/auth/Authentication';
@@ -37,14 +40,18 @@ const router = createBrowserRouter([
           {
             index: true,
             element: <HomePage />,
+            loader: postsLoader,
           },
           {
             path: ':postId',
+            id: 'post-detail',
             element: <PostDetailPage />,
+            loader: postDetailLoader,
           },
           {
             path: 'profile',
-            element: <div>Profile</div>,
+            element: <ProfilePage />,
+            loader: profileLoader,
           },
           {
             path: 'notification',
